Created initial pass at recurring task regex
