+++ src/main/kotlin/NeuralLinkPlugin.kt
@@ -1,5 +1,4 @@
 import event.FileModifiedEvent
-import grammar.RecurringGrammar
 import service.SettingsService
 import service.TaskService
 
@@ -27,13 +26,14 @@ class NeuralLinkPlugin(override var app: App, override var manifest: PluginManif
         // Add Settings tab
         addSettingTab(NeuralLinkPluginSettingsTab(app, this, settingsService, state))
         console.log("NeuralLinkPlugin onload()")
+
+        testStuff()
     }
 
     private fun testStuff() {
-        val grammar = RecurringGrammar()
-        console.log(grammar.parse("monthly!: 3"))
-        console.log(grammar.parse("daily: 1"))
-        console.log(grammar.parse("jan: 24"))
+        console.log(taskService.parseRecurring("monthly!: 3"))
+        console.log(taskService.parseRecurring("daily: 1"))
+        console.log(taskService.parseRecurring("jan: 24"))
     }
 
     override fun onunload() {
