+++ src/main/kotlin/service/TaskService.kt
@@ -6,6 +6,29 @@ import kotlin.js.Date
 @JsExport
 class TaskService {
     private val dueDateRegex = Regex("""@due\(([0-9\-T:]*)\)""")
+    private val recurRegex = Regex("""(daily|weekly|monthly|yearly|month|jan|feb|mar|apr|may|jun|jul|aug|sep|oct|nov|dec)([!]?): ([0-9]{1,2})""")
+
+    fun parseRecurring(recurText: String) : RecurItem {
+        val matches = recurRegex.find(recurText)
+        console.log("matches: ", matches)
+        if (matches?.groupValues == null) {
+            return RecurItem("ERROR", "", false, 0)
+        }
+
+        var index = 1
+        val type = matches?.groupValues?.get(index++)!!
+        val fromComplete = matches.groupValues[index++] == "!"
+        val amount = matches.groupValues[index]!!.toInt()
+
+        var recurType: String
+        if (type in listOf("daily", "weekly", "monthly", "yearly")) {
+            recurType = "SPAN"
+        } else {
+            recurType = "SPECIFIC"
+        }
+
+        return RecurItem(recurType, type, fromComplete, amount)
+    }
 
     /**
      * Gets a date from the task string tagged with the given tag.
@@ -25,3 +48,7 @@ class TaskService {
         } ?: return null
     }
 }
+
+@OptIn(ExperimentalJsExport::class)
+@JsExport
+data class RecurItem(val type: String, val span: String, val fromComplete: Boolean, val amount: Int)
